hit a roadblock
summarization will be harder, both the abstractive and extractive methods ive tried do not convey important information

i think removing filler, then working with what the result is is the best way to go. removing everything that isnt aligned with what the person is saying about what they are going to do, but just filler.